fix: Maximum call stack size exceeded with array.push (Bug: last_indexed_block reset to Zero after reaching the seed blocks #300)
When indexing contracts with a big number of events the array.push method will throw an error.

When you use spread operator all items of the source array are stored in the stack as arguments list, so having a large number of items (~ > 100K) will cause the this stack size exceed.

The most simple work-around is to manually push all items one by one.